+++ app/src/main/java/com/runnect/runnect/data/repository/CourseRepositoryImpl.kt
@@ -12,7 +12,7 @@ import com.runnect.runnect.data.dto.response.ResponsePostMyHistory
 import com.runnect.runnect.data.dto.response.ResponsePutMyDrawCourse
 import com.runnect.runnect.data.dto.response.ResponsePostDiscoverUpload
 import com.runnect.runnect.data.dto.response.ResponsePostScrap
-import com.runnect.runnect.data.mapToResult
+import com.runnect.runnect.data.network.mapToResult
 import com.runnect.runnect.data.source.remote.RemoteCourseDataSource
 import com.runnect.runnect.domain.common.Result
 import com.runnect.runnect.domain.entity.DiscoverSearchCourse
