+++ app/src/main/java/com/runnect/runnect/data/service/CourseV2Service.kt
@@ -1,6 +1,6 @@
 package com.runnect.runnect.data.service
 
-import com.runnect.runnect.data.ApiResult
+import com.runnect.runnect.data.network.ApiResult
 import com.runnect.runnect.data.dto.response.ResponseGetDiscoverMarathon
 import kotlinx.coroutines.flow.Flow
 import retrofit2.http.GET
