+++ app/src/main/java/com/runnect/runnect/di/RetrofitModule.kt
@@ -3,8 +3,8 @@ package com.runnect.runnect.di
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import com.runnect.runnect.BuildConfig
 import com.runnect.runnect.application.ApplicationClass
-import com.runnect.runnect.data.calladapter.flow.FlowCallAdapterFactory
-import com.runnect.runnect.data.interceptor.ResponseInterceptor
+import com.runnect.runnect.data.network.calladapter.FlowCallAdapterFactory
+import com.runnect.runnect.data.network.interceptor.ResponseInterceptor
 import com.runnect.runnect.data.service.*
 import com.runnect.runnect.data.repository.*
 import com.runnect.runnect.data.source.remote.*
